% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_functions.R
\name{add_stats_and_save}
\alias{add_stats_and_save}
\title{Add Nucleosome signal, TSSEnrichment, Blacklist and peak counts, and mitochondiral genes}
\usage{
add_stats_and_save(seur_obj, qs_save_path = "data/seur_stats.qs")
}
\arguments{
\item{seur_obj}{}
}
\value{
The seurat object with added stats
}
\description{
Add Nucleosome signal, TSSEnrichment, Blacklist and peak counts, and mitochondiral genes
}
