% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyStudyPackage-package.R
\docType{package}
\name{familyStudyPackage-package}
\alias{familyStudyPackage}
\alias{familyStudyPackage-package}
\title{familyStudyPackage: Analysis of familyStudyRNA.}
\description{
Analysis of familyStudyRNA
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/koefoeden/familyStudyPackage}
  \item Report bugs at \url{https://github.com/koefoeden/familySt“udyPackage/issues}
}

}
\author{
\strong{Maintainer}: Thomas Koefoed Gade \email{thomas.koefoed@sund.ku.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
