% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_clinical_data}
\alias{get_clinical_data}
\title{Load clinical data from DNA methylation paper}
\usage{
get_clinical_data()
}
\value{
A tibble of the wide metadata
}
\description{
Load clinical data from DNA methylation paper
}
