% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_combined_pheno_data}
\alias{get_combined_pheno_data}
\title{Get a phenotypic meta data table for the RNA seq analysis}
\usage{
get_combined_pheno_data()
}
\value{
Wide, dedupped data for each sample that we have RNAseq data for
}
\description{
Get a phenotypic meta data table for the RNA seq analysis
}
