% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_DGE_data}
\alias{get_DGE_data}
\title{Performs the test for each coefficient in the linear model}
\usage{
get_DGE_data(y_all, gene_names_type = "NCBI")
}
\arguments{
\item{gene_names_type}{"NCBI" or "ENSEMBL"}

\item{fit_all}{the fitted DGE list}
}
\description{
Performs the test for each coefficient in the linear model
}
