% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_interesting_genes_df}
\alias{get_interesting_genes_df}
\title{Get dataframe of interesting genes}
\usage{
get_interesting_genes_df()
}
\value{
Vector of gene names
}
\description{
Get dataframe of interesting genes
}
