% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_long_term_exercise_induced_genes}
\alias{get_long_term_exercise_induced_genes}
\title{Function to get genes from a table https://www-nature-com.ep.fjernadgang.kb.dk/articles/s41467-019-13869-w#Sec1
that includes exercise induced genes. Filtering for the long term}
\usage{
get_long_term_exercise_induced_genes()
}
\value{
The chararcter vector of genes
}
\description{
Function to get genes from a table https://www-nature-com.ep.fjernadgang.kb.dk/articles/s41467-019-13869-w#Sec1
that includes exercise induced genes. Filtering for the long term
}
