% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_MD_data_for_sample}
\alias{get_MD_data_for_sample}
\title{Plot MDS for a given dimensions and color it}
\usage{
get_MD_data_for_sample(
  object,
  column,
  xlab = "Average log CPM (this sample and others)",
  ylab = "log-ratio (this sample vs others)",
  prior.count = 3,
  ...
)
}
\arguments{
\item{object}{The DGElist object with transcript/gene counts and sample information
encoded in the y$sample object}

\item{y}{The DGElist object with transcript/gene counts and sample information}

\item{dims}{A numeric vector of the two dimensions of the MDS plot}

\item{color_by}{Character vector of the colour to color the points and labels by,
encoded in the y$sample object}
}
\value{
The ggplot-object with title, labels, and appropriate color

A long tibble with two columns: mean and diff values
}
\description{
Plot MDS for a given dimensions and color it
}
\examples{
ggplot_mds_repel()
Get MD data for a single sample

get_MD_data_for_all_samples()
}
