% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_PCA_data}
\alias{get_PCA_data}
\title{Get PCA data from Plink PCA output by loading file and formatting it}
\usage{
get_PCA_data(eigenvec_data = "data-raw/fam_study_only_genotyped.eigenvec")
}
\arguments{
\item{eigenvec_data}{The path to the PLINK PCA output}
}
\value{
Formatted PCA data with sourceId as identifying column
}
\description{
Get PCA data from Plink PCA output by loading file and formatting it
}
