% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pheno_data}
\alias{get_pheno_data}
\title{Load Phnomics platform long phenotypic data file}
\usage{
get_pheno_data()
}
\value{
A tibble of the long phenotypic data
}
\description{
Load Phnomics platform long phenotypic data file
}
