% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pheno_data_w_genetics}
\alias{get_pheno_data_w_genetics}
\title{Get pheno data w genetic info}
\usage{
get_pheno_data_w_genetics(results_dir)
}
\arguments{
\item{results_dir}{Path to a RNA seq analysis data folder}
}
\value{
Returns a tibble of all samples that have RNA data with their
clinical characteristics, RNA and genetic data
}
\description{
Get pheno data w genetic info
}
