% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pheno_data_wide}
\alias{get_pheno_data_wide}
\title{Loads and converts the long phenomics metadata to wide format, fixing datatypes}
\usage{
get_pheno_data_wide()
}
\value{
A wide phenomics metadata table
}
\description{
Loads and converts the long phenomics metadata to wide format, fixing datatypes
}
