% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_rna_counts}
\alias{get_rna_counts}
\title{Load RNA counts from the .rds file}
\usage{
get_rna_counts()
}
\value{
Count matrix for each sample with NCBI gene names
}
\description{
Load RNA counts from the .rds file
}
