% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_whole_DGE_list_by_tissue}
\alias{get_whole_DGE_list_by_tissue}
\title{Creates a fitted DGE list from tissue vector and model formula stirng}
\usage{
get_whole_DGE_list_by_tissue(tissue = "muscle")
}
\arguments{
\item{tissue}{the tissue to keept}

\item{formula}{the formula for the model as a string}
}
\value{
the fitted DGE list
}
\description{
Creates a fitted DGE list from tissue vector and model formula stirng
}
