% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ggplot_volcano}
\alias{ggplot_volcano}
\title{Make a volcano plot using ggplot}
\usage{
ggplot_volcano(df, type = "edgeR")
}
\arguments{
\item{The}{gene results from EdgeR or voom}

\item{Designate}{if the results are from EdgeR or voom}
}
\value{
The ggplot object
}
\description{
Make a volcano plot using ggplot
}
\examples{
ggplot_volcano()
}
