% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_salmon_metadata}
\alias{make_salmon_metadata}
\title{Construct a corresponding meta data tibble from the salmon count matrix}
\usage{
make_salmon_metadata(rna_counts)
}
\arguments{
\item{rna_counts}{}
}
\value{
Meta data tibble with Sample.ID, Family, Family-num, Tissue and type.
}
\description{
Construct a corresponding meta data tibble from the salmon count matrix
}
