% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{render_markdownfile_single_tisue_updated}
\alias{render_markdownfile_single_tisue_updated}
\title{Renders a markdown template file, filtering for a given tissue and outliers}
\usage{
render_markdownfile_single_tisue_updated(
  markdown_path,
  tissue,
  exclude,
  name = ""
)
}
\arguments{
\item{markdown_path}{The absolute path to the markdown fle}

\item{tissue}{Character vector of length 1, indicating the tissue to keep,
as present in the metadata sheet for the project}

\item{exclude}{Character vector of length 1, indicating the Sample IDS to
remove seperated by commas, as present in the metadata sheet for the project}
}
\value{
Returns nothing, but creates the corresponding .html files with descriptive names
}
\description{
Renders a markdown template file, filtering for a given tissue and outliers
}
