% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_for_phenotype}
\alias{search_for_phenotype}
\title{Search for phenotype and get counts in the dataset}
\usage{
search_for_phenotype(my_string)
}
\arguments{
\item{my_string}{Part of the phenotype to search for}
}
\value{
Returns the phenotype if it exists a long with the number of counts
}
\description{
Search for phenotype and get counts in the dataset
}
