% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{top_50_most_common_phenotypes}
\alias{top_50_most_common_phenotypes}
\title{Get top 50 most common phenotypes}
\usage{
top_50_most_common_phenotypes()
}
\value{
The top 50 most common phenotypes along with the counts
}
\description{
Get top 50 most common phenotypes
}
