% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_gene_info_df}
\alias{write_gene_info_df}
\title{Writes a gene info data frame to RDS with the ENSEMBL ID, HGNC symbol and gene description}
\usage{
write_gene_info_df(genes_name_type = "NCBI", organism = "human")
}
\arguments{
\item{genes_name_type}{"NCBI" or "ENSEMBL"}

\item{organism}{"human", "rat" or "mouse"}
}
\value{
Nothing.
}
\description{
Writes a gene info data frame to RDS with the ENSEMBL ID, HGNC symbol and gene description
}
