% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyStudyPackage-package.R
\docType{package}
\name{familyStudyPackage-package}
\alias{familyStudyPackage}
\alias{familyStudyPackage-package}
\title{familyStudyPackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/koefoeden/familyStudyPackage}
  \item Report bugs at \url{https://github.com/koefoeden/familyStudyPackage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
