% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ggplot_volcano_updated}
\alias{ggplot_volcano_updated}
\title{Make a volcano plot using ggplot}
\usage{
ggplot_volcano_updated(df, genes = "all")
}
\arguments{
\item{df}{gene results from EdgeR or voom}

\item{genes}{character vector, "all" for plotting all genes,
otherwise only plot vector of genes}
}
\value{
The ggplot object
}
\description{
Make a volcano plot using ggplot
}
